pyproject.toml
[project]
name = "lic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "numpy==1.26.4",
    "pandas==2.1.4",
    "pyarrow==14.0.2",
    "streamlit==1.32.0",
    "altair==4.2.2",
    "openpyxl==3.1.2",
    "python-docx==1.1.2",
    "ipykernel>=7.2.0",
]

[tool.setuptools.packages.find]
where = ["src"]